refactor(Trip): refactor Trip structure
