+++ pago-backend/src/main/java/tw/pago/pagobackend/dao/TripDao.java
@@ -3,9 +3,9 @@ package tw.pago.pagobackend.dao;
 import java.sql.SQLException;
 import java.util.List;
 
-import tw.pago.pagobackend.entity.Trip;
+import tw.pago.pagobackend.model.Trip;
 
-public interface TripDAO {
+public interface TripDao {
     public Trip getById(Integer tripId) throws SQLException;
 
     public List<Trip> findAll() throws SQLException;
